requirements.txt
click==4.0
sphinx-rtd-theme==0.1.8
requests==2.7.0
subprocess32==3.2.6
logger==1.4
pyyaml==3.11
fabric==1.10.2
pyvbox==0.2.2
python-vagrant==0.5.9
cuisine==0.7.10
python-keystoneclient==1.3.2
python-neutronclient==2.3.12
prettytable==0.7.2
ipaddress==1.0.14
beautifulsoup4==4.4.0
jenkinsapi==0.2.28
python-gerrit==0.0.1
